requirements.txt
discord-py
python-dotenv
matplotlib
requests
aiohttp
supabase